requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
jinja2==3.1.6
aiofiles==24.1.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
openai==1.59.5
httpx==0.28.1
python-dotenv==1.0.1
python-magic==0.4.27
